# Chunk iterators: Optimize histogram memory allocations