--- a/pkg/querier/batch/chunk.go
+++ b/pkg/querier/batch/chunk.go
@@ -7,7 +7,9 @@ package batch
 
 import (
 	"github.com/prometheus/common/model"
+	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
+	"github.com/prometheus/prometheus/util/zeropool"
 
 	"github.com/grafana/mimir/pkg/storage/chunk"
 )
@@ -17,16 +19,17 @@ import (
 type chunkIterator struct {
 	chunk GenericChunk
 	it    chunk.Iterator
-	batch *chunk.Batch
+	batch chunk.Batch
+
+	hPool  *zeropool.Pool[*histogram.Histogram]
+	fhPool *zeropool.Pool[*histogram.FloatHistogram]
 }
 
 func (i *chunkIterator) reset(chunk GenericChunk) {
 	i.chunk = chunk
 	i.it = chunk.Iterator(i.it)
-	if i.batch != nil {
-		i.batch.Length = 0
-		i.batch.Index = 0
-	}
+	i.batch.Length = 0
+	i.batch.Index = 0
 }
 
 // Seek advances the iterator forward to the value at or after
@@ -40,20 +43,17 @@ func (i *chunkIterator) Seek(t int64, size int) chunkenc.ValueType {
 
 	// If the seek is to the middle of the current batch, and size fits, we can
 	// shortcut.
-	if i.batch != nil {
-		if i.batch.Length > 0 && t >= i.batch.Timestamps[0] && t <= i.batch.Timestamps[i.batch.Length-1] {
-			i.batch.Index = 0
-			for i.batch.Index < i.batch.Length && t > i.batch.Timestamps[i.batch.Index] {
-				i.batch.Index++
-			}
-			if i.batch.Index+size < i.batch.Length {
-				return i.batch.ValueType
-			}
+	if i.batch.Length > 0 && t >= i.batch.Timestamps[0] && t <= i.batch.Timestamps[i.batch.Length-1] {
+		i.batch.Index = 0
+		for i.batch.Index < i.batch.Length && t > i.batch.Timestamps[i.batch.Index] {
+			i.batch.Index++
+		}
+		if i.batch.Index+size < i.batch.Length {
+			return i.batch.ValueType
 		}
 	}
-
 	if typ := i.it.FindAtOrAfter(model.Time(t)); typ != chunkenc.ValNone {
-		i.batch = i.it.Batch(size, typ, i.batch)
+		i.batch = i.it.Batch(size, typ, i.hPool, i.fhPool)
 		if i.batch.Length > 0 {
 			return typ
 		}
@@ -63,7 +63,7 @@ func (i *chunkIterator) Seek(t int64, size int) chunkenc.ValueType {
 
 func (i *chunkIterator) Next(size int) chunkenc.ValueType {
 	if typ := i.it.Scan(); typ != chunkenc.ValNone {
-		i.batch = i.it.Batch(size, typ, i.batch)
+		i.batch = i.it.Batch(size, typ, i.hPool, i.fhPool)
 		if i.batch.Length > 0 {
 			return typ
 		}
@@ -76,7 +76,7 @@ func (i *chunkIterator) AtTime() int64 {
 }
 
 func (i *chunkIterator) Batch() chunk.Batch {
-	return *i.batch
+	return i.batch
 }
 
 func (i *chunkIterator) Err() error {