--- a/pkg/querier/batch/chunk.go
+++ b/pkg/querier/batch/chunk.go
@@ -17,14 +17,16 @@ import (
 type chunkIterator struct {
 	chunk GenericChunk
 	it    chunk.Iterator
-	batch chunk.Batch
+	batch *chunk.Batch
 }
 
 func (i *chunkIterator) reset(chunk GenericChunk) {
 	i.chunk = chunk
 	i.it = chunk.Iterator(i.it)
-	i.batch.Length = 0
-	i.batch.Index = 0
+	if i.batch != nil {
+		i.batch.Length = 0
+		i.batch.Index = 0
+	}
 }
 
 // Seek advances the iterator forward to the value at or after
@@ -38,18 +40,20 @@ func (i *chunkIterator) Seek(t int64, size int) chunkenc.ValueType {
 
 	// If the seek is to the middle of the current batch, and size fits, we can
 	// shortcut.
-	if i.batch.Length > 0 && t >= i.batch.Timestamps[0] && t <= i.batch.Timestamps[i.batch.Length-1] {
-		i.batch.Index = 0
-		for i.batch.Index < i.batch.Length && t > i.batch.Timestamps[i.batch.Index] {
-			i.batch.Index++
-		}
-		if i.batch.Index+size < i.batch.Length {
-			return i.batch.ValueType
+	if i.batch != nil {
+		if i.batch.Length > 0 && t >= i.batch.Timestamps[0] && t <= i.batch.Timestamps[i.batch.Length-1] {
+			i.batch.Index = 0
+			for i.batch.Index < i.batch.Length && t > i.batch.Timestamps[i.batch.Index] {
+				i.batch.Index++
+			}
+			if i.batch.Index+size < i.batch.Length {
+				return i.batch.ValueType
+			}
 		}
 	}
 
 	if typ := i.it.FindAtOrAfter(model.Time(t)); typ != chunkenc.ValNone {
-		i.batch = i.it.Batch(size, typ)
+		i.batch = i.it.Batch(size, typ, i.batch)
 		if i.batch.Length > 0 {
 			return typ
 		}
@@ -59,7 +63,7 @@ func (i *chunkIterator) Seek(t int64, size int) chunkenc.ValueType {
 
 func (i *chunkIterator) Next(size int) chunkenc.ValueType {
 	if typ := i.it.Scan(); typ != chunkenc.ValNone {
-		i.batch = i.it.Batch(size, typ)
+		i.batch = i.it.Batch(size, typ, i.batch)
 		if i.batch.Length > 0 {
 			return typ
 		}
@@ -72,7 +76,7 @@ func (i *chunkIterator) AtTime() int64 {
 }
 
 func (i *chunkIterator) Batch() chunk.Batch {
-	return i.batch
+	return *i.batch
 }
 
 func (i *chunkIterator) Err() error {