--- a/pkg/querier/batch/merge.go
+++ b/pkg/querier/batch/merge.go
@@ -8,8 +8,11 @@ package batch
 import (
 	"container/heap"
 	"sort"
+	"unsafe"
 
+	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
+	"github.com/prometheus/prometheus/util/zeropool"
 
 	"github.com/grafana/mimir/pkg/storage/chunk"
 )
@@ -22,8 +25,10 @@ type mergeIterator struct {
 	batches batchStream
 
 	// Buffers to merge in.
-	batchesBuf   batchStream
-	nextBatchBuf [1]chunk.Batch
+	batchesBuf          batchStream
+	nextBatchBuf        [1]chunk.Batch
+	hPointerValuesPool  zeropool.Pool[unsafe.Pointer]
+	fhPointerValuesPool zeropool.Pool[unsafe.Pointer]
 
 	currErr error
 }
@@ -35,6 +40,8 @@ func newMergeIterator(it iterator, cs []GenericChunk) *mergeIterator {
 		c.currErr = nil
 	} else {
 		c = &mergeIterator{}
+		c.hPointerValuesPool = zeropool.New(func() unsafe.Pointer { return unsafe.Pointer(&histogram.Histogram{}) })
+		c.fhPointerValuesPool = zeropool.New(func() unsafe.Pointer { return unsafe.Pointer(&histogram.FloatHistogram{}) })
 	}
 
 	css := partitionChunks(cs)
@@ -70,6 +77,18 @@ func newMergeIterator(it iterator, cs []GenericChunk) *mergeIterator {
 	return c
 }
 
+func (c *mergeIterator) putPointerValuesToThePool(b chunk.Batch) {
+	if b.ValueType == chunkenc.ValHistogram || b.ValueType == chunkenc.ValFloatHistogram {
+		for i := 0; i < b.Length; i++ {
+			if b.ValueType == chunkenc.ValHistogram {
+				c.hPointerValuesPool.Put(b.PointerValues[i])
+			} else {
+				c.fhPointerValuesPool.Put(b.PointerValues[i])
+			}
+		}
+	}
+}
+
 func (c *mergeIterator) Seek(t int64, size int) chunkenc.ValueType {
 
 	// Optimisation to see if the seek is within our current caches batches.
@@ -83,6 +102,9 @@ found:
 			}
 			break found
 		}
+		// The first batch is not needed anymore, so we put pointers to its Histograms/FloatHistograms
+		// to the pool in order to reuse them.
+		c.putPointerValuesToThePool(c.batches[0])
 		copy(c.batches, c.batches[1:])
 		c.batches = c.batches[:len(c.batches)-1]
 	}
@@ -114,6 +136,9 @@ found:
 func (c *mergeIterator) Next(size int) chunkenc.ValueType {
 	// Pop the last built batch in a way that doesn't extend the slice.
 	if len(c.batches) > 0 {
+		// The first batch is not needed anymore, so we put pointers to its Histograms/FloatHistograms
+		// to the pool in order to reuse them.
+		c.putPointerValuesToThePool(c.batches[0])
 		copy(c.batches, c.batches[1:])
 		c.batches = c.batches[:len(c.batches)-1]
 	}
@@ -131,7 +156,7 @@ func (c *mergeIterator) buildNextBatch(size int) chunkenc.ValueType {
 	// is before all iterators next entry.
 	for len(c.h) > 0 && (len(c.batches) == 0 || c.nextBatchEndTime() >= c.h[0].AtTime()) {
 		c.nextBatchBuf[0] = c.h[0].Batch()
-		c.batchesBuf = mergeStreams(c.batches, c.nextBatchBuf[:], c.batchesBuf, size)
+		c.batchesBuf = mergeStreams(c.batches, c.nextBatchBuf[:], c.batchesBuf, size, false, true, &c.hPointerValuesPool, &c.fhPointerValuesPool)
 		c.batches = append(c.batches[:0], c.batchesBuf...)
 
 		if c.h[0].Next(size) != chunkenc.ValNone {