… pkg/querier/batch/batch_test.go
… pkg/querier/batch/chunk_test.go
… pkg/querier/batch/merge_test.go
… pkg/querier/batch/non_overlapping_test.go
… pkg/querier/batch/stream_test.go
… pkg/querier/querier_test.go
… pkg/storage/chunk/chunk.go
… pkg/storage/chunk/chunk_test.go
… pkg/storage/chunk/prometheus_chunk.go