#### What this PR does

Prometheus introduced a possibility to optionally reuse the memory area of the given histogram or float histogram to reduce allocation (https://github.com/prometheus/prometheus/pull/13340). 

In #7274  we adapted Mimir's `chunk.Iterator` interface in such a way that its `AtHistogram()` and `AtFloatHistogram()` methods follow the signature of the corresponding methods of Prometheus' `chunkenc.Iterator`, i.e., they now also contain optional arguments of type `*histogram.Histogram` and `*histogram.FloatHistogram` that allow memory allocation optimization. 

### Changes in `batch.iteratorAdapter` 
In #7219 we adapter signatures of `batch.iteratorAdapter.AtHistogram()` and `batch.iteratorAdapter.AtFloatHistogram()` by adding the optional argument of type `*histogram.Histogram` and `*histogram.FloatHistogram` needed for memory allocation optimization, but these arguments were not used. This PR changes the two implementations as it is explained in their documentations. Namely, the PR changes them from
```
	...
	// AtHistogram implements chunkenc.Iterator. It does not copy the underlying histogram as that optimization is left to the caller / higher level.
	func (a *iteratorAdapter) AtHistogram(h *histogram.Histogram) (int64, *histogram.Histogram) {
		...
	}
	...
	// AtFloatHistogram implements chunkenc.Iterator. It does not copy the underlying histogram as that optimization is left to the caller / higher level.
	func (a *iteratorAdapter) AtFloatHistogram(fh *histogram.FloatHistogram) (int64, *histogram.FloatHistogram) {
		...
	}
```
to
```
	...
	// AtHistogram implements chunkenc.Iterator. It copies and returns the underlying histogram.
	// If a pointer to a histogram is passed as parameter, the underlying histogram is copied there.
	// Otherwise, a new histogram is created.
	func (a *iteratorAdapter) AtHistogram(h *histogram.Histogram) (int64, *histogram.Histogram) {
		...
	}
	...
	// AtFloatHistogram implements chunkenc.Iterator. It copies and returns the underlying float histogram.
	// If a pointer to a float histogram is passed as parameter, the underlying float histogram is copied there.
	// Otherwise, a new float histogram is created.
	func (a *iteratorAdapter) AtFloatHistogram(fh *histogram.FloatHistogram) (int64, *histogram.FloatHistogram) {
		...
	}
```

### Changes in `chunk.Iterator` and  `chunk.prometheusChunkIterator`

In this PR we additionally modify `chunk.Iterator` interface by allowing its `Batch()` method to optimize memory allocations. The signature of the method is changed from 
```
	// Returns a batch of the provisded size; NB not idempotent!  Should only be called
	// once per Scan.
	Batch(size int, valueType chunkenc.ValueType) Batch
```
to
```
	// Batch returns a batch of the provisded size; NB not idempotent!  Should only be called
	// once per Scan.
	// When the optional *zeropool.Pool arguments hPool and fhPool are passed, they will be
	// used to optimize memory allocations for histogram.Histogram and histogram.FloatHistogram
	// objects.
	// For example, when creating a batch of chunkenc.ValHistogram or chunkenc.ValFloatHistogram
	// objects, the histogram.Histogram or histogram.FloatHistograms objects already present in
	// the hPool or fhPool pool will be used instead of creating new ones.
	Batch(size int, valueType chunkenc.ValueType, hPool *zeropool.Pool[*histogram.Histogram], fhPool *zeropool.Pool[*histogram.FloatHistogram]) Batch
```
Namely, `chunk.Iterator.Batch()` now accepts optional `*zeropool.Pool[*histogram.Histogram]` and `*zeropool.Pool[*histogram.FloatHistogram]` arguments that, when not nil, will avoid creation of new `histogram.Histogram` and `histogram.FloatHistogram` objects, if some already created objects of the same type already exist in the pools. The only known implementation of `chunk.Iterator`, `chunk.prometheusChunkIterator`, has been modified in such a way that it doesn't create new `histogram.Histogram` and `histogram.FloatHistogram` objects, when they can be reused.

### Changes in `batch.mergeIterator` 

Additionally, this PR modifies the implementation of `batch.mergeIterator` in such a way that it uses the benefits of the new `chunk.Iterator.Batch()` method, and further optimizes the memory allocations. Previously, new `histogram.Histogram` and `histogram.FloatHistogram` objects were created every time a new `Batch` was built by `batch.mergeIterator.buildNextBatch()`. With the present implementation, this situation changes: namely , when `batch.mergeIterator.build` is called, new `histogram.Histogram` and `histogram.FloatHistogram` will be created only if the pools (of type `zeropool.Pool`) storing pointers to unused `histogram.Histogram` and `histogram.FloatHistogram` objects are empty. Unused `histogram.Histogram` and `histogram.FloatHistogram` are the ones that get discarded on `batch.mergeIterator.Next()` and `batch.mergeIterator.Seek()`.

### Benchmark
```
                                                                                                                 │ old2602c.txt │             new2602c.txt             │
                                                                                                                 │    sec/op    │    sec/op     vs base                │
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-8              11.53m ±  3%   12.07m ±  5%   +4.69% (p=0.000 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusHistogramChunk-8        89.05m ± 36%   69.81m ±  4%  -21.60% (p=0.000 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusFloatHistogramChunk-8   81.34m ± 12%   70.21m ± 57%        ~ (p=0.123 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-8              43.31m ± 17%   22.50m ± 21%  -48.06% (p=0.000 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusHistogramChunk-8        244.0m ±  7%   104.3m ± 12%  -57.23% (p=0.000 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusFloatHistogramChunk-8   214.1m ±  7%   120.5m ± 12%  -43.69% (p=0.000 n=10)
geomean                                                                                                            75.75m         50.58m        -33.23%

                                                                                                                 │ old2602c.txt  │             new2602c.txt             │
                                                                                                                 │     B/op      │     B/op      vs base                │
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-8               8.585Mi ± 0%   8.585Mi ± 0%        ~ (p=0.796 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusHistogramChunk-8         75.81Mi ± 0%   33.09Mi ± 0%  -56.36% (p=0.000 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusFloatHistogramChunk-8    52.91Mi ± 0%   33.08Mi ± 0%  -37.48% (p=0.000 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-8               9.030Mi ± 0%   9.026Mi ± 0%   -0.05% (p=0.003 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusHistogramChunk-8        116.02Mi ± 0%   33.64Mi ± 0%  -71.00% (p=0.000 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusFloatHistogramChunk-8    93.12Mi ± 0%   33.62Mi ± 0%  -63.89% (p=0.000 n=10)
geomean                                                                                                             38.70Mi        21.40Mi       -44.72%

                                                                                                                 │ old2602c.txt │            new2602c.txt             │
                                                                                                                 │  allocs/op   │  allocs/op   vs base                │
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusXorChunk-8               3.376k ± 0%   3.377k ± 0%        ~ (p=0.221 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusHistogramChunk-8        1407.5k ± 0%   607.8k ± 0%  -56.82% (p=0.000 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_1_encoding:_PrometheusFloatHistogramChunk-8    907.5k ± 0%   607.7k ± 0%  -33.03% (p=0.000 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusXorChunk-8               9.395k ± 0%   9.395k ± 0%        ~ (p=0.865 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusHistogramChunk-8        2021.5k ± 0%   622.1k ± 0%  -69.23% (p=0.000 n=10)
QueryExecute/chunks:_1000_samples_per_chunk:_100_duplication_factor:_3_encoding:_PrometheusFloatHistogramChunk-8   1521.5k ± 0%   622.1k ± 0%  -59.11% (p=0.000 n=10)
geomean                                                                                                             223.5k        128.7k       -42.43%
```

#### Which issue(s) this PR fixes or relates to

Part of https://github.com/grafana/mimir/issues/7235

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.